Use external messages in tandem with custom hooks
Instead of having a single word that's configured as the action, take a
messageKey that then allows external users to have more control over the
language that shows up in the log files.

Also allow arbitrary extra parameters to then be added for use in that
message if needed.  This removes the need for the comment.

Issue #1: Show more information about who did what when